Add the ability to use existing GCloud installations when deploying

+++ deploy.sh
@@ -1,5 +1,27 @@
 #!/bin/bash
 
+
+usage() {
+  echo "Usage: $0 [--create-gcloud-installation]" 1>&2;
+  exit 1;
+}
+
+# =========================
+#    Read CLI Arguments.
+# =========================
+do_install_gcloud=false
+
+for var in "$@"
+do
+  if [[ "$var" == "--create-gcloud-installation" ]]; then
+    do_install_gcloud=true
+  else
+    usage
+    exit 1
+  fi
+done
+
+
 echo "=========================="
 echo "Installing dependencies..."
 echo "=========================="
@@ -9,10 +31,56 @@ echo " "
 echo "=================================="
 echo "Installing the Google Cloud SDK..."
 echo "=================================="
-rm -rf .GoogleCloud || { echo "Unable to remove old .GoogleCloud installation!" ; exit 1; }
-mkdir .GoogleCloud || { echo "Unable to create .GoogleCloud directory to install Google Cloud into!" ; exit 1; }
-curl https://sdk.cloud.google.com > .GoogleCloud/install.sh || { echo "Unable to download Google Cloud install script!" ; exit 1; }
-bash .GoogleCloud/install.sh --disable-prompts --install-dir=.GoogleCloud || { echo "Unable to install Google Cloud!" ; exit 1; }
+
+gcloud_command="gcloud"
+gsutil_command="gsutil"
+gcloud_installed=true
+
+# If we want to create our own installation of GCloud, then
+# we just ignore whether it has already been installed.
+if [ "$do_install_gcloud" = true ]; then
+  gcloud_installed=false
+fi
+if ! command -v "$gcloud_command" &> /dev/null; then
+  gcloud_installed=false
+fi
+if ! command -v "$gsutil_command" &> /dev/null; then
+  gcloud_installed=false
+fi
+
+if [ "$gcloud_installed" = true ]; then
+  echo " "
+  echo "Detected an existing external GCloud installation to use instead of installing."
+  echo "If this is not desired, then you may pass the \"--create-gcloud-installation\""
+  echo "option to this script."
+else
+  # Detect a previous installation of GCloud.
+  gcloud_command=".GoogleCloud/google-cloud-sdk/bin/gcloud"
+  gsutil_command=".GoogleCloud/google-cloud-sdk/bin/gsutil"
+  gcloud_installed=true
+
+  # If we want to clean the installation of GCloud, then
+  # we just ignore whether it has already been installed.
+  if ! command -v "$gcloud_command" &> /dev/null; then
+    gcloud_installed=false
+  fi
+  if ! command -v "$gsutil_command" &> /dev/null; then
+    gcloud_installed=false
+  fi
+
+  if [ "$gcloud_installed" = true ]; then
+    echo " "
+    echo "Detected a previous GCloud installation to use instead of installing."
+    echo "If this is not desired, then you may delete the \".GoogleCloud\""
+    echo "directory to re-install GCloud."
+  else
+    rm -rf .GoogleCloud || { echo "Unable to remove old .GoogleCloud installation!" ; exit 1; }
+    mkdir .GoogleCloud || { echo "Unable to create .GoogleCloud directory to install Google Cloud into!" ; exit 1; }
+    curl https://sdk.cloud.google.com > .GoogleCloud/install.sh || { echo "Unable to download Google Cloud install script!" ; exit 1; }
+    bash .GoogleCloud/install.sh --disable-prompts --install-dir=.GoogleCloud || { echo "Unable to install Google Cloud!" ; exit 1; }
+  fi
+fi
+
 
 echo " "
 echo "========================================================="
@@ -45,13 +113,13 @@ echo " "
 echo "==================================="
 echo "Authenticating with Google Cloud..."
 echo "==================================="
-.GoogleCloud/google-cloud-sdk/bin/gcloud auth login || { echo "Authenticating Google Cloud failed!" ; exit 1; }
+"$gcloud_command" auth login || { echo "Authenticating Google Cloud failed!" ; exit 1; }
 
 echo " "
 echo "============================"
 echo "Initialising Google Cloud..."
 echo "============================"
-.GoogleCloud/google-cloud-sdk/bin/gcloud init --skip-diagnostics || { echo "Initialising Google Cloud failed!" ; exit 1; }
+"$gcloud_command" init --skip-diagnostics || { echo "Initialising Google Cloud failed!" ; exit 1; }
 
 echo " "
 echo "==============================================="
@@ -60,7 +128,7 @@ echo "==============================================="
 echo " "
 echo "Please enter the URL of your project as a .appspot.com domain (e.g. misinformation-game.appspot.com):"
 read -r URL || exit 1
-.GoogleCloud/google-cloud-sdk/bin/gsutil cors set src/config/cors.json "gs://$URL" || { echo "Uploading CORS settings failed!" ; exit 1; }
+"$gsutil_command" cors set src/config/cors.json "gs://$URL" || { echo "Uploading CORS settings failed!" ; exit 1; }
 
 echo " "
 echo "======================"
